requirements.txt
# Core
fastapi==0.122.0
uvicorn==0.38.0
python-dotenv==1.1.1
pydantic==2.11.10
pydantic-settings==2.10.1
python-multipart==0.0.20

# LLM
openai==1.83.0

# RAG
faiss-cpu
sentence-transformers
numpy==2.3.5

# Agents
crewai==1.7.2
mcp==1.16.0

# Rate Limiting
slowapi==0.1.9

# GitHub cloning
gitpython

# PDF support
pypdf==6.4.0
pdfplumber==0.11.8

# HTTP
requests==2.32.5
httpx==0.28.1

# Streaming
sse-starlette==3.1.2

# UI
streamlit
